+++ src/Form/FieldSubmit.js
@@ -1,4 +1,4 @@
-import classNames from 'classnames';
+import classNames from 'classnames/dedupe';
 import React from 'react';
 
 class FieldSubmit extends React.Component {
@@ -10,18 +10,14 @@ class FieldSubmit extends React.Component {
     } = this.props;
 
     let rowClass = classNames(
-      `column-${columnWidth}`,
+      `form-row-element column-${columnWidth}`,
       formElementClass
     );
 
-    let buttonClassSet = classNames({
-      [buttonClass]: !!buttonClass
-    });
-
     return (
       <div className={rowClass}>
-        <div className={formGroupClass}>
-          <div className={buttonClassSet}
+        <div className={classNames(formGroupClass)}>
+          <div className={classNames(buttonClass)}
             onClick={handleSubmit}>
             {buttonText}
           </div>
@@ -34,9 +30,14 @@ class FieldSubmit extends React.Component {
 
 FieldSubmit.defaultProps = {
   buttonText: 'Submit',
-  formElementClass: 'form-row-element',
   handleSubmit: function () {}
 };
+
+let classPropType = React.PropTypes.oneOfType([
+  React.PropTypes.array,
+  React.PropTypes.object,
+  React.PropTypes.string
+]);
 
 FieldSubmit.propTypes = {
   // Optional number of columns to take up of the grid
@@ -50,9 +51,9 @@ FieldSubmit.propTypes = {
   buttonText: React.PropTypes.string,
 
   // Classes
-  buttonClass: React.PropTypes.string,
-  formGroupClass: React.PropTypes.string,
-  formElementClass: React.PropTypes.string
+  buttonClass: classPropType,
+  formGroupClass: classPropType,
+  formElementClass: classPropType
 };
 
 module.exports = FieldSubmit;